eric-esoa-shared-spring-security/src/main/java/com/ericsson/bos/so/shared/spring/security/HttpAuthenticationEntryPoint.java
/*******************************************************************************
 * COPYRIGHT Ericsson 2023
 *
 *
 *
 * The copyright to the computer program(s) herein is the property of
 *
 * Ericsson Inc. The programs may be used and/or copied only with written
 *
 * permission from Ericsson Inc. or in accordance with the terms and
 *
 * conditions stipulated in the agreement/contract under which the
 *
 * program(s) have been supplied.
 ******************************************************************************/
package com.ericsson.bos.so.shared.spring.security;

import java.io.IOException;


import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.web.AuthenticationEntryPoint;
import org.springframework.stereotype.Component;

/**
 * HttpAuthenticationEntryPoint - configures to set response  as not authorised
 */
@Component
public class HttpAuthenticationEntryPoint implements AuthenticationEntryPoint {

    /**
     *
     * commence - set response of request to unauthorized
     *
     * @param httpServletRequest - the http request sent
     * @param httpServletResponse - the http response received
     * @param e - AuthenticationException that was thrown
     * @throws IOException
     */
    @Override
    public void commence(HttpServletRequest httpServletRequest,
                         HttpServletResponse httpServletResponse, AuthenticationException e) throws IOException {

        httpServletResponse.sendError(HttpServletResponse.SC_UNAUTHORIZED, e.getMessage());
    }
}
